Cast attributes to ints

+++ source/Ship.cpp
@@ -2678,12 +2678,12 @@ int Ship::TakeDamage(const Projectile &projectile, bool isBlast)
 	shields -= shieldDamage * shieldFraction;
 	if(shieldDamage && !isDisabled)
 	{
-		int disabledDelay = attributes.Get("depleted shield delay");
-		shieldDelay = max(shieldDelay, (shields <= 0. && disabledDelay) ? disabledDelay : attributes.Get("shield delay"));
+		int disabledDelay = static_cast<int>((attributes.Get("depleted shield delay"));
+		shieldDelay = max(shieldDelay, (shields <= 0. && disabledDelay) ? disabledDelay : static_cast<int>(attributes.Get("shield delay")));
 	}
 	hull -= hullDamage * (1. - shieldFraction);
 	if(hullDamage && !isDisabled)
-		hullDelay = max(hullDelay, attributes.Get("repair delay"));
+		hullDelay = max(hullDelay, static_cast<int>(attributes.Get("repair delay")));
 	// For the following damage types, the total effect depends on how much is
 	// "leaking" through the shields.
 	double leakage = (1. - .5 * shieldFraction);
@@ -2709,7 +2709,7 @@ int Ship::TakeDamage(const Projectile &projectile, bool isBlast)
 	if(!wasDisabled && isDisabled)
 	{
 		type |= ShipEvent::DISABLE;
-		hullDelay = max(hullDelay, attributes.Get("disabled repair delay"));
+		hullDelay = max(hullDelay, static_cast<int>(attributes.Get("disabled repair delay")));
 	}
 	if(!wasDestroyed && IsDestroyed())
 		type |= ShipEvent::DESTROY;
